+++ lib/src/core/config/constants/cache_invalidation_constants.dart
@@ -1,6 +1,6 @@
-import '../enums/invalidation_type.dart';
+import '../../../modules/data_cache/domain/enums/invalidation_type.dart';
 
-final class CacheConstants {
+final class CacheInvalidationConstants {
   static const defaultInvalidationType = InvalidationType.ttl;
   static const maxTtlDuration = Duration(days: 1);
 }
