+++ lib/src/modules/data_cache/domain/entities/cache_entity.dart
@@ -36,9 +36,6 @@ class CacheEntity<T extends Object> {
   /// Timestamp indicating when the cache entity was last updated (nullable).
   final DateTime? updatedAt;
 
-  /// Constructs a `CacheEntity` instance.
-  ///
-  /// The `id`, `data`, `invalidationType`, `createdAt`, and `endAt` parameters are required.
   const CacheEntity({
     required this.id,
     required this.data,
@@ -48,9 +45,6 @@ class CacheEntity<T extends Object> {
     required this.endAt,
   });
 
-  /// Constructs a `CacheEntity` instance with generated timestamps.
-  ///
-  /// The `id`, `data`, `invalidationType`, and `expireMaxDuration` parameters are required.
   factory CacheEntity._generate({
     required String id,
     required T data,
@@ -68,9 +62,6 @@ class CacheEntity<T extends Object> {
     );
   }
 
-  /// Constructs a `CacheEntity` instance from a data transfer object (DTO).
-  ///
-  /// The `dto` parameter is required.
   factory CacheEntity.toSave(SaveCacheDTO<T> dto) {
     return CacheEntity._generate(
       id: dto.key,
@@ -90,21 +81,6 @@ class CacheEntity<T extends Object> {
     );
   }
 
-  /// Creates a copy of this cache entity with optional new values for its fields.
-  ///
-  /// This method provides a way to clone the current instance while possibly
-  /// changing some of its properties. It's useful for modifying instances of
-  /// immutable classes.
-  ///
-  /// - Parameters:
-  ///   - `id`: An optional new unique identifier for the cache entity.
-  ///   - `data`: An optional new data of type `T` for the cache entity.
-  ///   - `invalidationType`: An optional new invalidation type for the cache entity.
-  ///   - `createdAt`: An optional new creation date and time for the cache entity.
-  ///   - `endAt`: An optional new expiration date and time for the cache entity.
-  ///   - `updatedAt`: An optional new update date and time for the cache entity.
-  ///
-  /// - Returns: A new `CacheEntity<T>` instance with updated fields as provided.
   CacheEntity<T> copyWith({
     String? id,
     T? data,
@@ -123,9 +99,6 @@ class CacheEntity<T extends Object> {
     );
   }
 
-  /// Equality comparison for `CacheEntity`.
-  ///
-  /// Two cache entities are considered equal if their unique identifiers `id` are equal.
   @override
   bool operator ==(Object other) {
     if (identical(this, other)) return true;
@@ -133,9 +106,6 @@ class CacheEntity<T extends Object> {
     return other is CacheEntity<T> && other.id == id;
   }
 
-  /// Hash code generation for `CacheEntity`.
-  ///
-  /// Returns the hash code based on the unique identifier `id` of the cache entity.
   @override
   int get hashCode => id.hashCode;
 }
