+++ lib/src/modules/data_cache/domain/services/size_analyzer_service/cache_size_analyzer_service.dart
