+++ lib/src/modules/data_cache/domain/usecases/clear_cache_usecase.dart
@@ -2,10 +2,16 @@ import '../../../../core/core.dart';
 
 import '../repositories/i_cache_repository.dart';
 
+/// Abstract interface for the ClearCache use case.
 abstract interface class ClearCacheUsecase {
+  /// Executes the clear cache operation.
+  ///
+  /// Returns an [AsyncEither] containing either an [AutoCacheError]
+  /// or a [Unit] upon completion.
   AsyncEither<AutoCacheError, Unit> execute();
 }
 
+/// Implementation of the [ClearCacheUsecase].
 class ClearCache implements ClearCacheUsecase {
   final ICacheRepository _repository;
