+++ test/src/modules/data_cache/domain/usecases/get_cache_usecase_test.dart
@@ -2,7 +2,7 @@ import 'package:auto_cache_manager/src/core/core.dart';
 import 'package:auto_cache_manager/src/modules/data_cache/domain/dtos/get_cache_dto.dart';
 import 'package:auto_cache_manager/src/modules/data_cache/domain/entities/cache_entity.dart';
 import 'package:auto_cache_manager/src/modules/data_cache/domain/enums/invalidation_type.dart';
-import 'package:auto_cache_manager/src/modules/data_cache/domain/services/invalidation/invalidation_cache_context.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/domain/services/invalidation_service/invalidation_cache_context.dart';
 import 'package:auto_cache_manager/src/modules/data_cache/domain/usecases/get_cache_usecase.dart';
 import 'package:auto_cache_manager/src/modules/data_cache/infra/repositories/cache_repository.dart';
 import 'package:flutter_test/flutter_test.dart';
